tests:  check if ctags is available
TestBuilder assumed that ctags is not available during testing. With
this change the test works whether ctags is present or not.

Since Zoekt doesn't strictly require ctags we should handle both cases.
On CI, however, we should run the tests with ctags, which I will tackle
in a separate PR.